+++ src/VaultSharp/V1/SystemBackend/HealthStatus.cs
@@ -77,7 +77,7 @@ namespace VaultSharp.V1.SystemBackend
         /// <value>
         /// The name of the cluster.
         /// </value>
-        [JsonProperty("cluster_name")]
+        [JsonProperty("cluster_name", NullValueHandling = NullValueHandling.Ignore)]
         public string ClusterName { get; set; }
 
         /// <summary>
@@ -86,7 +86,7 @@ namespace VaultSharp.V1.SystemBackend
         /// <value>
         /// The cluster identifier.
         /// </value>
-        [JsonProperty("cluster_id")]
+        [JsonProperty("cluster_id", NullValueHandling = NullValueHandling.Ignore)]
         public string ClusterId { get; set; }
 
         /// <summary>