+++ src/VaultSharp/V1/SystemBackend/LogVerbosityLevel.cs
@@ -1,0 +1,14 @@
+﻿namespace VaultSharp.V1.SystemBackend
+{
+    /// <summary>
+    /// The LogVerbosityLevel.
+    /// </summary>
+    public enum LogVerbosityLevel
+    {
+        trace,
+        debug,
+        info,
+        warn,
+        error
+    }
+}